# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
        "/organizations/building/{buildingId}": {
            "get": {
                "summary": "Получить организации по ID здания",
                "operationId": "3a494125abed56884ac37696169e00b7",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "ID of the building",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "ООО “Рога и Копыта”"
                                            },
                                            "building_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "phone_numbers": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "number": {
                                                            "type": "string",
                                                            "example": "2-222-222"
                                                        },
                                                        "organization_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "activities": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Еда"
                                                        },
                                                        "parent_id": {
                                                            "type": "integer",
                                                            "example": null,
                                                            "nullable": true
                                                        },
                                                        "pivot": {
                                                            "properties": {
                                                                "organization_id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "activity_id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Building not found"
                    }
                }
            }
        },
        "/organizations/activity/{activityId}": {
            "get": {
                "summary": "Получить организации по ID деятельности",
                "operationId": "d073284996b36f32412f10a54ea23cc0",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "ООО “Рога и Копыта”"
                                            },
                                            "building_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "phone_numbers": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "number": {
                                                            "type": "string",
                                                            "example": "2-222-222"
                                                        },
                                                        "organization_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "activities": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Еда"
                                                        },
                                                        "parent_id": {
                                                            "type": "integer",
                                                            "example": null,
                                                            "nullable": true
                                                        },
                                                        "pivot": {
                                                            "properties": {
                                                                "organization_id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "activity_id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Building not found"
                    }
                }
            }
        },
        "/organizations/radius": {
            "get": {
                "summary": "Получить организации по радиусу",
                "operationId": "4d519641f83dac666becd1782df58867",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Широта точки",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Долгота точки",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус в километрах",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "ООО “Рога и Копыта”"
                                            },
                                            "building_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "phone_numbers": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "number": {
                                                            "type": "string",
                                                            "example": "2-222-222"
                                                        },
                                                        "organization_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "activities": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Еда"
                                                        },
                                                        "parent_id": {
                                                            "type": "integer",
                                                            "example": null,
                                                            "nullable": true
                                                        },
                                                        "pivot": {
                                                            "properties": {
                                                                "organization_id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "activity_id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Организации не найдены"
                    }
                }
            }
        },
        "/organizations/{id}": {
            "get": {
                "summary": "Получить организацию по ID",
                "operationId": "3b641b1670623b572d4c510217881de0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "ООО “Рога и Копыта”"
                                        },
                                        "building_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "phone_numbers": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "number": {
                                                        "type": "string",
                                                        "example": "2-222-222"
                                                    },
                                                    "organization_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "activities": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Еда"
                                                    },
                                                    "parent_id": {
                                                        "type": "integer",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "pivot": {
                                                        "properties": {
                                                            "organization_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "activity_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Организация не найдена"
                    }
                }
            }
        }
    }
}